MW-SO_1.0.7_batchCode_issue_updated

+++ MWSO/app/src/main/java/com/techsmith/mw_so/retail_utils/PAYDETAIL.java
@@ -4,7 +4,7 @@ public class PAYDETAIL {
 
     public String PAYTYPE;
 
-    public int AMOUNT;
+    public float AMOUNT;
     public Double CARDAMOUNT;
 
     public String CARDNAME;
