+++ MWSO/app/src/main/java/com/techsmith/mw_so/retail_utils/PAYSUMMARY.java
@@ -1,6 +1,6 @@
 package com.techsmith.mw_so.retail_utils;
 
 public class PAYSUMMARY {
-    public int BILLAMOUNT;
-    public int PAIDAMOUNT;
+    public float BILLAMOUNT;
+    public float PAIDAMOUNT;
 }
