+++ MWSO/app/src/main/java/com/techsmith/mw_so/retail_utils/Summary.java
@@ -10,7 +10,7 @@ public class Summary {
     public String CCEID;
     public int SUMMARYDISC;
     public double ROUNDOFF;
-    public int NETAMOUNT;
+    public Float NETAMOUNT;
     public int NOOFITEMS;
     public String DOCGUID;
 }
